VB-1698 - updates to dtos.
